Update traffic lights duration
In order to help in creating a green wave, traffic lights duration need
to take account of the full length of the road, so the last light is
synchronised with the first one. Now duration of red light is set as a
percentage of the duration of full green/red light cycle.